+++ R/dbplyr.R
@@ -24,7 +24,6 @@
 #' @importFrom dplyr db_explain
 #' @importFrom dplyr db_has_table
 #' @importFrom dplyr db_insert_into
-#' @importFrom dplyr db_list_tables
 #' @importFrom dplyr db_query_fields
 #' @importFrom dplyr db_query_rows
 #' @importFrom dplyr db_rollback