+++ tests/testthat/test-backend-postgres.R
@@ -64,3 +64,9 @@ test_that("can explain", {
   db <- copy_to_test("postgres", data.frame(x = 1:3))
   expect_snapshot(db %>% mutate(y = x + 1) %>% explain())
 })
+
+test_that("can overwrite temp tables", {
+  src <- src_test("postgres")
+  copy_to(src, mtcars, "mtcars", overwrite = TRUE)
+  expect_error(copy_to(src, mtcars, "mtcars", overwrite = TRUE), NA)
+})